Update Reggie
*Tidy Execute method and found source of QueryParam bug
*Split different modules into different files
